authenticate sensordata

+++ iotedge-api/Endpoints/SensorData.py
@@ -15,21 +15,34 @@ parser.add_argument('ConnectionString', type=str, required=False)
 
 class SensorData(Resource):
 
-    #  Get all SensorData by owner
-    # @jwt_required
-    # def get(self):
-    #     storage = AzureTableStorage()
-    #     table_service = storage.get_table()
-    #     verify_jwt_in_request()
-    #     args = parser.parse_args()
-    #     filter = "OwnerId eq '{0}' and ConnectionString eq {1}".format(get_jwt_claims()["id"], )
-    #     filter = "PartitionKey eq '{0}'".format(args["ConnectionString"])
-    #     rows = table_service.query_entities('sensordata', filter=filter)
-    #     for row in rows:
-    #         row["Timestamp"] = row["Timestamp"].isoformat()
-    #     return {"message": "success", "sensordata": list(rows), "uri": request.base_url}, 200
+    #Get all SensorData by owner
+    @jwt_required
+    def get(self):
+        storage = AzureTableStorage()
+        table_service = storage.get_table()
+        verify_jwt_in_request()
+        args = parser.parse_args()
+        filter = "OwnerId eq '{0}'".format(get_jwt_claims()["id"])
+
+        connectionstrings = []
+        sensors = table_service.query_entities('sensors', filter=filter)
+        for sensor in sensors:
+            connectionstrings.append(sensor["ConnectionString"])
+
+        sensordatas = []
+        for connectionstring in connectionstrings:
+            filter = "PartitionKey eq '{0}'".format(connectionstring)
+            sensordata = table_service.query_entities('sensordata', filter=filter)
+            if len(list(sensordatas)) > 0:
+                for point in sensordata:
+                    sensordatas.append(point)
+
+        for row in sensordatas:
+            row["Timestamp"] = row["Timestamp"].isoformat()
+        return {"message": "success", "sensordata": sensordatas, "uri": request.base_url}, 200
 
     # Post new SensorData
+
     @jwt_required
     def post(self):
         storage = AzureTableStorage()
@@ -41,27 +54,38 @@ class SensorData(Resource):
         sensordata_table = table_service.query_entities('rulers', filter=filter)
         sensordata_table = list(sensordata_table)[0]
 
-        try:
-            sensors_fields = {
-                "PartitionKey": args["ConnectionString"].replace("'", ";"),
-                "RowKey": str(sensordata_table["NewId"]),
-                "Datavalue": args["Datavalue"].replace("'", ";"),
-            }
-        except:
-            return {"message": "fill all data"}, 400
+        filter = "ConnectionString eq '{0}'".format(args["ConnectionString"])
+        sensor = list(table_service.query_entities('sensors', filter=filter))[0]
+
+        if(len(sensor) > 0):
+            if sensor["OwnerId"] == get_jwt_claims()["id"]:
+                try:
+                    sensors_fields = {
+                        "PartitionKey": args["ConnectionString"].replace("'", ";"),
+                        "RowKey": str(sensordata_table["NewId"]),
+                        "Datavalue": args["Datavalue"].replace("'", ";"),
+                    }
+                except:
+                    return {"message": "fill all data"}, 400
+
+                table_service.insert_entity('sensors', sensors_fields)
+                ruler_sensors = {"PartitionKey": sensordata_table['PartitionKey'], "RowKey": sensordata_table['RowKey'],
+                                 "NewId": sensordata_table["NewId"] + 1, "Size": sensordata_table["Size"] + 1}
+                table_service.update_entity('rulers', ruler_sensors)
+
+                return {"message": "success", "sensordata": sensors_fields}, 200
+            else:
+                return {"message": "not your sensor"}, 400
+        else:
+            return {"message": "sensor not found"}, 400
 
 
-        table_service.insert_entity('sensors', sensors_fields)
-        ruler_sensors = {"PartitionKey": sensordata_table['PartitionKey'], "RowKey": sensordata_table['RowKey'],
-                         "NewId": sensordata_table["NewId"] + 1, "Size": sensordata_table["Size"] + 1}
-        table_service.update_entity('rulers', ruler_sensors)
 
-        return {"message": "success", "sensordata": sensors_fields}, 200
 
 
 class SingleSensorData(Resource):
 
-    # Get SensorData by id
+    # Get SensorData by id and partitionkey
     @jwt_required
     def get(self, partitionkey, id):
         storage = AzureTableStorage()
@@ -79,16 +103,21 @@ class SingleSensorData(Resource):
         else:
             return {"message": "error id not found"}, 400
 
-        filter = "PartitionKey eq '{0}' and {1} eq '{2}'".format(partitionkey, searcher, specification)
+        filter = "OwnerId eq {0} and PartitionKey eq {1}".format(get_jwt_claims()[0], partitionkey)
+        sensor = table_service.query_entities('sensors', filter=filter)
+        if len(list(sensor)) > 0:
+            filter = "PartitionKey eq '{0}' and {1} eq '{2}'".format(partitionkey, searcher, specification)
+            sensordata = table_service.query_entities('sensordata', filter=filter)
+            if len(list(sensordata)) > 0:
+                sensordata = list(sensordata)[0]
+            else:
+                return {"message": "error data not found"}, 400
 
-        sensordata = table_service.query_entities('sensordata', filter=filter)
-        if len(list(sensordata)) > 0:
-            sensordata = list(sensordata)[0]
+            sensordata["Timestamp"] = sensordata["Timestamp"].isoformat()
+            return {"message": "success", "sensordata": sensordata, "uri": request.base_url}, 200
         else:
-            return {"message": "error data not found"}, 400
+            return {"message": "not your sensordata"}
 
-        sensordata["Timestamp"] = sensordata["Timestamp"].isoformat()
-        return {"message": "success", "sensordata": sensordata, "uri": request.base_url}, 200
 
     # Update SensorData by id
     @jwt_required
@@ -108,26 +137,31 @@ class SingleSensorData(Resource):
         else:
             return {"message": "error device not found"}
 
-        filter = "PartitionKey eq '{0}' and {1} eq '{2}'".format(partitionkey, searcher, specification)
+        filter = "OwnerId eq {0} and PartitionKey eq {1}".format(get_jwt_claims()["id"], partitionkey)
+        sensor = table_service.query_entities('sensors', filter=filter)
+        if len(list(sensor)) > 0:
+            filter = "PartitionKey eq '{0}' and {1} eq '{2}'".format(partitionkey, searcher, specification)
 
-        sensordata = table_service.query_entities('sensordata', filter=filter)
-        if len(list(sensordata)) > 0:
-            sensordata = list(sensordata)[0]
+            sensordata = table_service.query_entities('sensordata', filter=filter)
+            if len(list(sensordata)) > 0:
+                sensordata = list(sensordata)[0]
+            else:
+                return {"message": "error device not found"}
+
+            try:
+                sensordata["PartitionKey"] = args["ConnectionString"].replace("'", ";")
+                sensordata["Name"] = args["Name"].replace("'", ";")
+                sensordata["Datavalue"] = args["Datavalue"].replace("'", ";")
+                del sensordata["etag"]
+            except:
+                return {"message": "fill data"}, 400
+
+            table_service.update_entity('sensordata', sensordata)
+
+            sensordata["Timestamp"] = sensordata["Timestamp"].isoformat()
+            return {"message": "success", "sensordata": sensordata, "uri": request.base_url}, 200
         else:
-            return {"message": "error device not found"}
-
-        try:
-            sensordata["PartitionKey"] = args["ConnectionString"].replace("'", ";")
-            sensordata["Name"] = args["Name"].replace("'", ";")
-            sensordata["Datavalue"] = args["Datavalue"].replace("'", ";")
-            del sensordata["etag"]
-        except:
-            return {"message": "fill data"}, 400
-
-        table_service.update_entity('sensordata', sensordata)
-
-        sensordata["Timestamp"] = sensordata["Timestamp"].isoformat()
-        return {"message": "success", "sensordata": sensordata, "uri": request.base_url}, 200
+            return {"message": "not your sensordata"}
 
     # Delete SensorData by id
     @jwt_required
@@ -147,40 +181,49 @@ class SingleSensorData(Resource):
         else:
             return {"message": "error device not found"}, 400
 
-        filter = "PartitionKey eq '{0}' and {1} eq '{2}'".format(partitionkey, searcher, specification)
+        filter = "OwnerId eq {0} and PartitionKey eq {1}".format(get_jwt_claims()["id"], partitionkey)
+        sensor = table_service.query_entities('sensors', filter=filter)
+        if len(list(sensor)) > 0:
+            filter = "PartitionKey eq '{0}' and {1} eq '{2}'".format(partitionkey, searcher, specification)
 
-        sensordata = table_service.query_entities('sensordata', filter=filter)
-        if len(list(sensordata)) > 0:
-            sensordata = list(sensordata)[0]
+            sensordata = table_service.query_entities('sensordata', filter=filter)
+            if len(list(sensordata)) > 0:
+                sensordata = list(sensordata)[0]
+            else:
+                return {"message": "error device not found"}, 400
+
+            table_service.delete_entity('sensordata', sensordata["PartitionKey"], sensordata["RowKey"])
+
+            return {"message": "success deleted sensordata {}".format(sensordata["Name"])}, 200
         else:
-            return {"message": "error device not found"}, 400
-
-        table_service.delete_entity('sensordata', sensordata["PartitionKey"], sensordata["RowKey"])
-
-        return {"message": "success deleted sensordata {}".format(sensordata["Name"])}, 200
-
+            return {"message": "not your sensordata"}
 
 class GetSensorSensorData(Resource):
 
     # Get all SensorData from ConnectionString
     @jwt_required
-    def get(self, id):
+    def get(self, id, partitionkey):
         storage = AzureTableStorage()
         table_service = storage.get_table()
         verify_jwt_in_request()
 
-        filter = "OwnerId eq '{0}' and RowKey eq '{1}'".format(get_jwt_claims()["id"], id.replace("'", ";"))
+        filter = "OwnerId eq '{0}' and RowKey eq '{1}' and PartitionKey '{2}'".format(get_jwt_claims()["id"], id.replace("'", ";"), partitionkey.replace("'", ";"))
 
-        sensordata = table_service.query_entities('sensors', filter=filter)
-        if len(list(sensordata)) > 0:
-            sensordata = list(sensordata)[0]
+        sensor = table_service.query_entities('sensors', filter=filter)
+        if len(list(sensor)) > 0:
+            sensor = list(sensor)[0]
         else:
             return {"message": "error data not found"}, 400
-        connectionstring = sensordata["ConnectionString"]
-
-        filter = "PartitionKey eq {0}".format(connectionstring)
+        connectionstring = sensor["ConnectionString"]
+        filter = "PartitionKey eq '{0}'".format(connectionstring)
         points = table_service.query_entities('sensordata', filter=filter)
-        for point in points:
+
+        if len(list(points)) > 0:
+            sensordata = list(points)
+        else:
+            return {"message": "error data not found"}, 400
+
+        for point in sensordata:
             point["Timestamp"] = point["Timestamp"].isoformat()
 
         return {"message": "success", "sensordata": sensordata, "uri": request.base_url}, 200
